Added premake support, imgui docking and project reorganization

+++ RandomEngine/src/Platform/Windows/WindowsWindow.cpp
@@ -1,6 +1,7 @@
 #include "REPCH.h"
 	
 #include <glad/glad.h>
+#include <GLFW/glfw3.h>
 
 #include "WindowsWindow.h"
 #include "RandomEngine/Events/ApplicationEvent.h"
@@ -128,14 +129,14 @@ namespace RandomEngine {
 		glfwSetScrollCallback(_window, [](GLFWwindow* window, double xOffset, double yOffset)
 			{
 				WindowData& data = *(WindowData*)glfwGetWindowUserPointer(window);
-				MouseScrolledEvent event(xOffset, yOffset);
+				MouseScrolledEvent event((float)xOffset, (float)yOffset);
 				data.EventCallback(event);
 			});
 
 		glfwSetCursorPosCallback(_window, [](GLFWwindow* window, double xPos, double yPos)
 			{
 				WindowData& data = *(WindowData*)glfwGetWindowUserPointer(window);
-				MouseMovedEvent event(xPos, yPos);
+				MouseMovedEvent event((float)xPos, (float)yPos);
 				data.EventCallback(event);
 			});
 	}
